Move assets to static folder
