+++ docs/STYLE_GUIDE.md
@@ -36,72 +36,6 @@ To set these tools up with VSCode:
 - Prettier (easy, install the VSCode extension and enable formatOnSave in userSettings)
 - [ESLint](https://github.com/airbnb/javascript/issues/1589#issuecomment-344097023)
 
-## Directory structure:
-
-```
-src
-	assets
-	fonts
-	images
-	icons
-		index.js
-	components
-		Example
-			__tests__
-				Example.test.js
-			ExampleDependentComponent
-				__tests__
-					ExampleDependentComponent.test.js
-				index.js
-			index.js
-			styles.js
-	config
-	containers
-	e2e
-		...
-		Example.spec.js
-	handlers
-		AuthHandler
-			__tests__
-			index.js
-	reducers
-		appData
-			__tests__
-			index.js
-			initialState.js
-		index.js
-	sagas
-		analytics
-			logEvent
-				__tests__
-				index.js
-		index.js
-	scenes
-		pages
-			Home
-				__tests__
-				index.js
-		modals
-	services
-		analytics
-			logEvent
-				__tests__
-				index.js
-		index.js
-	store
-	styleConstants
-		index.js
-		dimensions
-			index.js
-	utils
-		app
-			createError
-				__tests__
-				indexjs
-	App.js
-	Router.js
-```
-
 ## Redux actions
 
 We use the [Flux Standard Action](https://github.com/redux-utilities/flux-standard-action) as a standard for our redux actions.
