+++ src/components/Lightbox/styles.js
@@ -1,6 +1,6 @@
 import { StyleSheet } from 'react-native';
 
-import styleConstants from '../../styleConstants';
+import styleConstants from '../../static/styleConstants';
 
 const styles = StyleSheet.create({
   wrapper: {
