+++ src/static/icons/index.js
