+++ src/static/styleConstants/colors/index.js
