+++ src/static/styleConstants/helpers/index.js
