+++ src/static/styleConstants/index.js
